+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Unreleased
 
+### Fixed
+
+- (#93) Do not assume that the first part of the album such as **Neon White OST 1 - The
+  Wicked Heart** is an artist if it contains a reference to a sound track.
+
 ## [0.24.2] 2026-04-12
 
 ### Fixed