+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ### Fixed
 
-- (#93) Do not assume that the first part of the album such as **Neon White OST 1 - The
-  Wicked Heart** is an artist if it contains a reference to a sound track.
+- `album`: 
+  - (#93) Do not assume that the first part of the album such as **Neon White OST 1 - The
+    Wicked Heart** is an artist if it contains a reference to a sound track.
+  - Do not parse remix name as a catalognum in track titles such as **What Is Love (Funk
+    Tribu Edit 2K23)**
 
 ## [0.24.2] 2026-04-12